Small code improvement
