+++ webapp/src/components/sidebar/sidebarBoardItem.tsx
@@ -38,6 +38,7 @@ import {getCurrentTeam} from '../../store/teams'
 import {Permission} from '../../constants'
 import PageMenu from '../pageMenu'
 import {Utils} from '../../utils'
+import SidebarPageItem from './sidebarPageItem'
 
 import AddIcon from '../../widgets/icons/add'
 import CloseIcon from '../../widgets/icons/close'
@@ -70,102 +71,6 @@ type Props = {
     draggedItemID?: string
     hideViews?: boolean
     showPage: (pageId: string, boardId: string) => void
-}
-
-type PageSidebarItemProps = {
-    page: Page
-    pages: Page[]
-    board: Board
-    currentPageId: string
-    showPage: (pageId: string, boardId: string) => void
-    showBoard: (boardId: string) => void
-    depth: number
-}
-
-const PageSidebarItem = (props: PageSidebarItemProps) => {
-    const intl = useIntl()
-    const {page, pages, board, currentPageId, depth} = props
-    const subpages = pages.filter((p) => p.parentId == page.id)
-
-    return (
-        <>
-            <div
-                key={page.id}
-                className={`SidebarBoardItem sidebar-page-item ${page.id === currentPageId ? 'active' : ''} depth-${depth}`}
-                onClick={() => props.showPage(page.id, board.id)}
-            >
-                {page.fields.icon || <CompassIcon icon='file-text-outline'/>}
-                <div
-                    className='octo-sidebar-title'
-                    title={page.title || intl.formatMessage({id: 'Sidebar.untitled-page', defaultMessage: '(Untitled Page)'})}
-                >
-                    {page.title || intl.formatMessage({id: 'Sidebar.untitled-page', defaultMessage: '(Untitled Page)'})}
-                </div>
-                <PageMenu
-                    pageId={page.id}
-                    onClickDelete={async () => {
-                        if (!page) {
-                            Utils.assertFailure()
-                            return
-                        }
-                        TelemetryClient.trackEvent(TelemetryCategory, TelemetryActions.DeletePage, {board: props.board.id, page: page.id})
-                        await mutator.deleteBlock(page, 'delete page')
-                        props.showBoard(page.boardId)
-                    }}
-                    onClickDuplicate={async () => {
-                        if (!page) {
-                            Utils.assertFailure()
-                            return
-                        }
-                        TelemetryClient.trackEvent(TelemetryCategory, TelemetryActions.DuplicatePage, {board: props.board.id, page: page.id})
-                        await mutator.duplicatePage(
-                            page.id,
-                            props.board.id,
-                            'duplicate page',
-                            async (newPageId: string) => {
-                                props.showPage(newPageId, page.boardId)
-                            },
-                            async () => {
-                                props.showPage(page.id, page.boardId)
-                            },
-                        )
-                    }}
-                    onClickAddSubpage={async () => {
-                        if (!page) {
-                            Utils.assertFailure()
-                            return
-                        }
-                        TelemetryClient.trackEvent(TelemetryCategory, TelemetryActions.CreateSubpage, {board: props.board.id, page: page.id})
-                        const subpage = createPage()
-                        subpage.parentId = page.id
-                        subpage.boardId = board.id
-                        await mutator.insertBlock(
-                            board.id,
-                            subpage,
-                            intl.formatMessage({id: 'Mutator.new-subpage', defaultMessage: 'new subpage'}),
-                            async (newBlock: Block) => {
-                                props.showPage(newBlock.id, board.id)
-                            },
-                            async () => {
-                                props.showPage(currentPageId, board.id)
-                            },
-                        )
-                    }}
-                />
-            </div>
-            {subpages.map((page: Page) => (
-                <PageSidebarItem
-                    depth={props.depth+1}
-                    page={page}
-                    pages={pages}
-                    board={board}
-                    currentPageId={currentPageId}
-                    showPage={props.showPage}
-                    showBoard={props.showBoard}
-                />
-            ))}
-        </>
-    )
 }
 
 const SidebarBoardItem = (props: Props) => {
@@ -469,7 +374,7 @@ const SidebarBoardItem = (props: Props) => {
                     ))}
 
                     {props.isActive && pages.filter((p) => p.parentId === board.id).map((page: Page) => (
-                        <PageSidebarItem
+                        <SidebarPageItem
                             page={page}
                             pages={pages}
                             board={board}
